Fixed: repositories and entoties package (rename)
