+++ src/main/java/entities/Evaluacion.java
@@ -1,4 +1,4 @@
-package entitys;
+package entities;
 
 import java.util.Date;
