+++ src/main/java/repositories/EvaluacionRepository.java
@@ -1,8 +1,8 @@
-package repositorys;
+package repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import entitys.Evaluacion;
+import entities.Evaluacion;
 
 public interface EvaluacionRepository extends JpaRepository<Evaluacion, Integer> {
